refactor: separate WorkspaceStartScreen in desktop and mobile
