+++ frontend/appflowy_flutter/lib/user/presentation/screens/workspace_start_screen/mobile_workspace_start_screen.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class MobileWorkspaceStartScreen extends StatelessWidget {
+  const MobileWorkspaceStartScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO(yijing): implement MobileWorkspaceStartScreen later
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('MobileWorkspaceStartScreen'),
+      ),
+      body: const Placeholder(),
+    );
+  }
+}
