+++ frontend/appflowy_flutter/lib/user/presentation/screens/workspace_start_screen/workspace_start_screen.dart
@@ -1,0 +1,47 @@
+import 'dart:io';
+
+import 'package:appflowy/generated/locale_keys.g.dart';
+import 'package:appflowy/startup/startup.dart';
+import 'package:appflowy/user/presentation/screens/workspace_start_screen/desktop_workspace_start_screen.dart';
+import 'package:appflowy/user/presentation/screens/workspace_start_screen/mobile_workspace_start_screen.dart';
+import 'package:appflowy/workspace/application/workspace/workspace_bloc.dart';
+import 'package:appflowy_backend/protobuf/flowy-user/user_profile.pb.dart';
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+class WorkspaceStartScreen extends StatelessWidget {
+  final UserProfilePB userProfile;
+  static const routeName = "/WorkspaceStartScreen";
+  const WorkspaceStartScreen({
+    Key? key,
+    required this.userProfile,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (_) => getIt<WorkspaceBloc>(param1: userProfile)
+        ..add(const WorkspaceEvent.initial()),
+      child: BlocBuilder<WorkspaceBloc, WorkspaceState>(
+        builder: (context, state) {
+          if (Platform.isAndroid || Platform.isIOS) {
+            return const MobileWorkspaceStartScreen();
+          }
+          return DesktopWorkspaceStartScreen(
+            workspaceState: state,
+          );
+        },
+      ),
+    );
+  }
+}
+
+void _onCreateNewWorkspace(BuildContext context) {
+  context.read<WorkspaceBloc>().add(
+        WorkspaceEvent.createWorkspace(
+          LocaleKeys.workspace_hint.tr(),
+          "",
+        ),
+      );
+}
